Add shader program classes
Shader_program is the base class where the program is created, but the
shaders for each program are held by each program type individually.
The Shaders are passed as const& to shared_ptr to allow several programs
to use them.
Attachment of shaders and linking of the program is only performed in
the derived classes.

Uniform updates will be added in the future and should preferably be
implemented in the base class.