Revamp Mesh_loader and Mesh class again
By introducing a Mesh_data struct holding the mesh data in vectors the
SegFault issue is solved but the data is also stored in a reasonable
way. This way meshes are not stored in the Mesh_loader but can be
created from Mesh_data which is stored there.

This improves access to Mesh VBOs.