Revamp Buffer handling in VAOs and Meshes
Previously all buffer objects were stored by value in std::vector of
their type. This caused calls to the destructor if one copy of the
Buffer was destroyed and therefore deleted its data from the GPU. To
prevent this from happening all buffer objects are now stored and passed
as shared_ptr. By doing this, the destructor will only be called if the
shared object is no longer used by any of its owners.